Add files via upload
The dataset is accessed as a part of Jose Portilla's udemy course. The dataset is about an Ecommerce company based in New York City that sells clothing online but they also have in-store style and clothing advice sessions. Customers come in to the store, have sessions/meetings with a personal stylist, then they can go home and order either on a mobile app or website for the clothes they want.